requirement.txt
joblib==0.13.2
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
python-mnist==0.6
scikit-learn==0.21.3
scipy==1.3.1
